Basic text rendering engine
Text rendering will use a pool of bitmaps which sizes are power of two.
When drawing text, we’ll measure it (using the assigned font), and if
our current bitmap is not big enough (or way too big), we’ll take a new
bitmap with the most appropriate size from the pool.