Text rendering- support auto fit and text wrap
Added support for various options of how to match the text with its
label:
No fitting- the default, the label will have its set size and the text
will be sized based on its font.
Text wrapping- the label will have its set size and the text will wrap
so it won’t get out of the width of the label, but without scaling.
Text to label- the label will have its set size and the text will be
scaled to fit the label.
Label to text- the text will be sized based on the font and the label
will be scaled to fit the text.