Component framework- refactoring
1. Moved the partial classes for objects, characters and UI controls
(the classes that are not automatically generated) back to their
original locations.
2. Changed the T4 templates to use ($SolutionDir) VS macro instead of
absolute path in my local machine. This will only work in visual
studio, however. A solution for MonoDevelop (and perhaps for the
automated builds) is still needed (as a workaround, put your absolute
path before generating new code from the templates and then revert back
to $SolutionDir).