Support room transitions
Resolves #15

The room transitions is a state machine which involves 3 entities:
Renderer loop, game loop and player’s change room component.

The workflow is as follows:
1. Player changes room, notifies the rendering loop and waits before
actually moving the player.
2. Rendering loop allows the transition to do stuff before the move.
3. When transition finished, rendering loop creates a screenshot of the
old room.
4. Player’s change room component wakes up and moves the player to the
new room.
5. Game loop is notified to set up the camera for the new room.
6. Rendering loop creates a screenshot of the new room and sends the
two screenshots to the room transition to play with.
7. When room transition is finished, rendering loop gives the new
display list to the room transition to play with.
8. When room transition is finished, we return back to normal.