Demo game: Fixed load image config for bottle item
We currently load the bottle image twice, once for the room and once
for the inventory. Since it has the same id we actually share the
texture. The room load used the “black is transparent” but the
inventory did not (accidentally). Since the room happened first this
was ok and we didn’t notice the issue). When we switched to async
loading the order changed and we started seeing the black artifacts on
the bottle, hence this fix.