Fixed OpenAL error when stopping a sound
Stopping a sound used to set a task as completed, but audio clip
registers to that task and releases the source, which causes the
blocking loop on play to return an invalid name on the source when
querying the sound status, as the source doesn’t exist.
Removed the task completion, the blocking loop is now the only one who
sets the task as completed, when it sees that the source stopped
playing.
Also added a missing error check when setting the panning.