Fix Win32Error crash (on Mac!) from GDI+
To repro (weird repro steps warning):
1. Run demo game on mac
2. Before doing any walking with the character, click Alt+G to open the
debug view
3. Select one of the GUI objects (for example, “FPS Label”
4. In the inspector, click on the slider and then use the arrow keys to
move the slider a little bit
5. Click on one of the properties so that the slider will lose focus
6. Now play with the arrow keys again which should move the character,
but if you change animations fast enough it would trigger enabling
pixel perfect twice in a row at the “same time” which triggers the bug.

This is because Bitmap.LockBits is called twice in a row without
unlocking in between which crashes GDI+:
https://github.com/mono/libgdiplus/blob/0c0592d09c7393fc418fa7f65f54e8b3
bcc14cf2/src/bitmap.c#L1960.
Added a static collection to save pixel perfect areas that were already
registered, and if such exists but the area does not exist yet, it
means an operation is already on its way and we ignore the second
command.