Text rendering- fixed caching bugs
1. Textbox text did not render well. The reason: we ignored text
visible changes to fire bounding box changes, but as we avoid rendering
to bitmap if text not visible, this effectively missed rendering newly
entered text if the text without caret was invisible. Fixed by
including text visible in the notifications.

2. Text on the hotspot label did not render well. When calling
GLText.SetProperties, it might affect properties that trigger a need to
re-prepare the bounding boxes again on the next tick. It will now
return a boolean which we’ll use to trigger the dirty flag.

3. Text on tree did not render well when cropped. Reason: we reused the
text bounding boxes for the crop, thinking it refreshes again on every
prepare, but because we’re not refreshing it on every tick now, we need
a new variable to store this.