bug fix: Stack layout didn't perform layout
Stack layout only performed layout changes when the bounding box with
children changes, but there are scenarios where a bounding box changes
for one of the children but it doesn’t change the external bounding box
with children (because the entire change is contained inside the outer
box), so we need to listen to events for all the children, there’s no
way around this.
Added “EntitiesListSubscriptions” class to listen to children events
and used it in the stack layout component, and also in the bounding box
with children component.