Implemented SpriteRenderComponent, sprite-based rendering of an entity
This adds ISpriteRenderComponent interface and its AGSSpriteRenderComponent implementation. SpriteRenderer is supposed to become a rendering "contractor" for 2D sprite-based components, such as ImageComponent or AnimationComponent.
SpriteRenderComponent stores a reference to ISpriteProvider, from which it receives "current sprite" to draw.
AGSAnimationComponent implements ISpriteProvider.

Moved Border and DebugDrawPivot properties from Animation to SpriteRender component.
Made GLImageRenderer refer to SpriteRenderComponent for drawing (instead of AnimationComponent). It is assumed, that parts of GLImageRenderer rendering function will be merged into SpriteRenderComponent on later stages.