Refactor scrollbars out of the contents panel
Having the scrollbars in the same panel as the contents created issues
as they were excluded from cropping in their panel (and therefore had
their crop self component removed) but included in cropping in the
parent panel (which had their crop self component added) which created
conflicts.
Creating a scrolling panel will now create a contents panel, and the
scrolling panel will be expanded to contain the scrollbars so they will
also not be in front of the contents.