Bug fix- race condition when walking between rooms
If the user managed to sneak in a walk instruction when switching
rooms, right after the character changes room (and position) the
instruction from the previous room might have slipped under the radar
and the character would then change positions to the old positions of
the old room. Added a specific room check to prevent that.