Fix font loading
The previous code assumed that when adding a font to private memory
font collection, the font will be added to the last index in the array.
Turns out (sigh) that it is not the case, so need to iterate the array
and keep tabs on families we already loaded in the past in order to
find the family we just added.