Remove Say and additional blocking APIs
Removed blocking APIs as those will not work as desired in some cases
(For example if you use “Say” in after fade in event, you will not
actually see it because it will block the event from completing). The
asynchronous counterparts do everything that the blocking APIs do but
will work in all scenarios.