Remove ILocation interface and rename Location to Position-
Removing the ILocation interface adds the ability to use the tuple shortcuts, and adds a possibility for future performance enhancements.
The abstraction over Z that AGSLocation gave before was moved to AGSTranslate (to keep the benefit of the previous design which allowed customizing this behavior- now to override the Z behavior one would provide a custom ITranslate instead of a custom ILocation).