Room- background color property
Added a background color property for room. Note this "feature", while desired on its own, is actually a temporary solution to a bug we get on windows when we clear the screen (invalid access exception). Still no solution to this bug but at least now by default we don't clear the screen on every tick.