Follow- fix game freeze when unfollowing
When unfollowing, the follow component issued a "stop walking" command from repeatedly execute. Unfortunately, because the follow component subscribed to the event asynchronously, the rest of the subscribers had to wait. If the walk component subscribed after that it never got to its repeatedly execute so it couldn't complete the stop walking command, causing the deadlock.
Fixed by converting on repeatedly execute into a blocking event so you won't be able to subscribe asynchronously to it again.